+++ components/gallery/GalleryItemTabsPanel/GalleryItemOffers.vue
@@ -79,7 +79,6 @@
 <script lang="ts" setup>
 import { NeoSecondaryButton, NeoTable, NeoTableColumn } from '@kodadot1/brick'
 import Identity from '@/components/identity/IdentityIndex.vue'
-import { getKSMUSD } from '@/utils/coingecko'
 import formatBalance from '@/utils/format/balance'
 import { formatSecondsToDuration } from '@/utils/format/time'
 import type { Offer, OfferResponse } from '@/components/bsx/Offer/types'
@@ -234,7 +233,7 @@ watch(
   async ([offersData, collectionData]) => {
     const nftPrice = collectionData?.collectionEntity?.nfts[0]?.price
     if (offersData?.offers) {
-      const ksmPrice = await getKSMUSD()
+      const ksmPrice = await getApproximatePriceOf('kusama')
       const floorPrice = formatPrice(nftPrice || '')
 
       offers.value = offersData.offers
