+++ utils/constants.ts
@@ -67,6 +67,7 @@ export const URLS = {
   },
   providers: {
     coingecko: 'https://api.coingecko.com/api/v3',
+    kodaprice: 'https://coinprice.kodadot.workers.dev/',
     ramp: 'https://ramp.network/buy/',
     cloudflare: 'https://cloudflare-ipfs.com/ipfs/',
     pinata: 'https://api.pinata.cloud/',
